requirements.txt
black==22.1.0
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
python-dotenv==0.19.2
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
